requirements.txt
Django>=6.0,<7
djangorestframework
django-cors-headers
pymysql
